Flesh out French Press Recipe Page, adjust img
I added all of the steps and descriptions to the body section of the
french press html file, and then found an image of a french press with
a coffee mug to use.
This image turned out to be too big so I adjusted the height and width
attributes in the <img> element to 500 for each so that it would fit
properly on the webpage.