Add practice CSS to index.html and french-press.
I practiced using three forms of CSS, (internal, external, and inline)
I used external with a link on my index.html page to change the header.
To do this, I had to change it from h1 to div, to practice with the div
tag.
Then for internal, I adjusted the p tag in my french-press page.
Finally for inline, I added a button to the bottom of the index.html,
because I had not known how to make a button, I just typed it into the
editor and followed the prompts for the properties that it was supposed
to have.
Definitely need more practice with buttons because I still don't
know their full functionality.